AVX-32997: Move token
