+++ api/login.py
@@ -1,6 +1,35 @@
+import json
+
 import requests
 
-from aviatrix_ha import get_api_token, AvxError, MASK
+from tools.string_utils import MASK
+from errors.exceptions import AvxError
+
+
+def get_api_token(ip_addr):
+    """ Get API token from controller. Older controllers that don't support it will not have this
+    API or endpoints. We return None in that scenario to be backkward compatible """
+    try:
+        data = requests.get(f'https://{ip_addr}/v2/api?action=get_api_token', verify=False)
+    except requests.exceptions.ConnectionError as err:
+        print("Can't connect to controller with elastic IP %s. %s" % (ip_addr, str(err)))
+        raise AvxError(str(err)) from err
+    buf = data.content
+    try:
+        out = json.loads(buf)
+    except ValueError:
+        print(f"Token is probably not supported. Reponse is {buf}")
+    else:
+        try:
+            token = out['results']['api_token']
+        except (ValueError, AttributeError, TypeError, KeyError) as err:
+            print(f"Getting token failed due to {err}")
+            print(f"Token is probably not supported. Reponse is {out}")
+        else:
+            print('Obtained token')
+            return token
+    print('Did not obtain token')
+    return None
 
 
 def login_to_controller(ip_addr, username, pwd):
