+++ tools/string_utils.py
@@ -1,0 +1,1 @@
+MASK = lambda input: input[0:5] + '*' * 15 if isinstance(input, str) else ''
